Add files via upload
The game shows a grid of 24 cards facing upside down. The user has to pick two cards, if the cards match each other, they are are discarded otherwise they are turned upside down again, and the user has to pick two other cards. The game ends when the user has found all the pairs.